+++ db/migrate/20181023150745_add_restrict_clipping_group_size_to_draws.rb
@@ -1,0 +1,6 @@
+class AddRestrictClippingGroupSizeToDraws < ActiveRecord::Migration[5.1]
+  def change
+    add_column :draws, :restrict_clipping_group_size,
+      :boolean, null: false, default: false
+  end
+end