+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20180611183201) do
+ActiveRecord::Schema.define(version: 20181023150745) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -87,6 +87,7 @@ ActiveRecord::Schema.define(version: 20180611183201) do
     t.date "locking_deadline"
     t.integer "suite_selection_mode", default: 0, null: false
     t.boolean "allow_clipping", default: false, null: false
+    t.boolean "restrict_clipping_group_size", default: false, null: false
   end
 
   create_table "groups", force: :cascade do |t|