+++ spec/factories/draws.rb
@@ -4,6 +4,7 @@ FactoryBot.define do
   factory :draw do
     name 'MyString'
     allow_clipping true
+    restrict_clipping_group_size false
 
     factory :draw_with_members do
       transient do