+++ spec/forms/new_clip_form_spec.rb
@@ -3,9 +3,10 @@
 require 'rails_helper'
 
 RSpec.describe NewClipForm do
-  # This will create a clip of length 2 by default or a clip of length 3
-  # if the value '1' is passed into the `add_self` variable
-  let(:draw) { create(:draw, allow_clipping: true) }
+  # This will create a clip of length 2 by default
+  let(:draw) do
+    create(:draw, allow_clipping: true, restrict_clipping_group_size: false)
+  end
   let(:groups) { create_pair(:group_from_draw, draw: draw) }
   let(:group_ids) { groups.map(&:id).map(&:to_s) }
   let(:group) { create(:group_from_draw, draw: draw) }
@@ -13,42 +14,35 @@ RSpec.describe NewClipForm do
 
   # Each shared 'success' example must define the following params in let blocks
   # user [User] a user with the role being tested
-  # add_self [String] if the user is a housing rep set it to '1' to add
-  #   them to the clip, '0' to make the clip without them. Set this to nil if
-  #   the user is not a housing rep.
   # size [Integer] the expected size of the created clip
   # mail_count [Integer] the expected amount of email invites sent out
   shared_examples 'success' do
     let(:param_hash) do
-      { draw_id: draw.id.to_s, group_ids: group_ids,
-        add_self: add_self }
+      { draw_id: draw.id.to_s, group_ids: group_ids }
     end
     let(:p) { instance_spy('ActionController::Parameters', to_h: param_hash) }
 
     it 'creates a clip and sets the redirect object correctly' do
-      r = described_class.new(admin: user.admin?, params: p).submit
+      r = described_class.new(role: user.role, params: p).submit
       expect(r[:redirect_object]).to be_instance_of(Clip)
     end
     it 'creates a clip of correct size' do
-      r = described_class.new(admin: user.admin?, params: p).submit
+      r = described_class.new(role: user.role, params: p).submit
       clip = r[:redirect_object]
       expect(clip.clip_memberships.length).to eq(size)
     end
     it 'returns a success flash message' do
-      r = described_class.new(admin: user.admin?, params: p).submit
+      r = described_class.new(role: user.role, params: p).submit
       expect(r[:msg]).to have_key(:success)
     end
     it 'sets :form_object to nil' do
-      r = described_class.new(admin: user.admin?, params: p).submit
+      r = described_class.new(role: user.role, params: p).submit
       expect(r[:form_object]).to eq(nil)
     end
   end
 
   # Each shared 'failure' example must define the following params in let blocks
   # user [User] a user with the role being tested
-  # add_self [Boolean] if the user is a housing rep set it to true to add
-  #   them to the clip, false to make the clip without them. Set this to nil if
-  #   the user is not a housing rep.
   # invalid_groups [Array<String>] group ids in an array that are too few in
   #   number to pass the size validations in the creator (can be nil or
   #   empty strings)
@@ -58,13 +52,13 @@ RSpec.describe NewClipForm do
       it 'sets the redirect object to nil' do
         param_hash = { draw_id: '', group_ids: group_ids }
         p = instance_spy('ActionController::Parameters', to_h: param_hash)
-        r = described_class.new(admin: user.admin?, params: p).submit
+        r = described_class.new(role: user.role, params: p).submit
         expect(r[:redirect_object]).to be_nil
       end
       it 'returns the form object' do
         param_hash = { draw_id: '', group_ids: group_ids }
         p = instance_spy('ActionController::Parameters', to_h: param_hash)
-        r = described_class.new(admin: user.admin?, params: p).submit
+        r = described_class.new(role: user.role, params: p).submit
         expect(r[:form_object]).to be_instance_of(described_class)
       end
     end
@@ -73,13 +67,13 @@ RSpec.describe NewClipForm do
       it 'sets the redirect object to nil' do
         param_hash = { draw_id: draw.id.to_s, group_ids: invalid_groups }
         p = instance_spy('ActionController::Parameters', to_h: param_hash)
-        r = described_class.new(admin: user.admin?, params: p).submit
+        r = described_class.new(role: user.role, params: p).submit
         expect(r[:redirect_object]).to be_nil
       end
       it 'returns the form object' do
         param_hash = { draw_id: draw.id.to_s, group_ids: invalid_groups }
         p = instance_spy('ActionController::Parameters', to_h: param_hash)
-        r = described_class.new(admin: user.admin?, params: p).submit
+        r = described_class.new(role: user.role, params: p).submit
         expect(r[:form_object]).to be_instance_of(described_class)
       end
     end
@@ -87,12 +81,12 @@ RSpec.describe NewClipForm do
     context 'invalid params are passed' do
       it 'sets the redirect object to nil' do
         p = instance_spy('ActionController::Parameters', to_h: {})
-        r = described_class.new(admin: user.admin?, params: p).submit
+        r = described_class.new(role: user.role, params: p).submit
         expect(r[:redirect_object]).to be_nil
       end
       it 'returns the form object' do
         p = instance_spy('ActionController::Parameters', to_h: {})
-        r = described_class.new(admin: user.admin?, params: p).submit
+        r = described_class.new(role: user.role, params: p).submit
         expect(r[:form_object]).to be_instance_of(described_class)
       end
     end
@@ -103,19 +97,19 @@ RSpec.describe NewClipForm do
       it 'sets the redirect object to nil' do
         param_hash = { draw_id: clipless_draw.id.to_s, group_ids: group_ids }
         p = instance_spy('ActionController::Parameters', to_h: param_hash)
-        r = described_class.new(admin: user.admin?, params: p).submit
+        r = described_class.new(role: user.role, params: p).submit
         expect(r[:redirect_object]).to be_nil
       end
       it 'returns the form object' do
         param_hash = { draw_id: clipless_draw.id.to_s, group_ids: group_ids }
         p = instance_spy('ActionController::Parameters', to_h: param_hash)
-        r = described_class.new(admin: user.admin?, params: p).submit
+        r = described_class.new(role: user.role, params: p).submit
         expect(r[:form_object]).to be_instance_of(described_class)
       end
       it 'returns the correct error message' do
         param_hash = { draw_id: clipless_draw.id.to_s, group_ids: group_ids }
         p = instance_spy('ActionController::Parameters', to_h: param_hash)
-        r = described_class.new(admin: user.admin?, params: p).submit
+        r = described_class.new(role: user.role, params: p).submit
         msg = 'This draw currently does not allow for clipping.'
         expect(r[:msg][:error]).to include(msg)
       end
@@ -126,14 +120,14 @@ RSpec.describe NewClipForm do
         param_hash = { draw_id: draw.id.to_s, group_ids: group_ids }
         p = instance_spy('ActionController::Parameters', to_h: param_hash)
         groups.first.clip = clip
-        r = described_class.new(admin: user.admin?, params: p).submit
+        r = described_class.new(role: user.role, params: p).submit
         expect(r[:redirect_object]).to be_nil
       end
       it 'returns the form object' do
         param_hash = { draw_id: draw.id.to_s, group_ids: group_ids }
         p = instance_spy('ActionController::Parameters', to_h: param_hash)
         groups.first.clip = clip
-        r = described_class.new(admin: user.admin?, params: p).submit
+        r = described_class.new(role: user.role, params: p).submit
         expect(r[:form_object]).to be_instance_of(described_class)
       end
     end
@@ -143,16 +137,41 @@ RSpec.describe NewClipForm do
         new_draw = create(:draw)
         param_hash = { draw_id: new_draw.id.to_s, group_ids: [[group.id]] }
         p = instance_spy('ActionController::Parameters', to_h: param_hash)
-        r = described_class.new(admin: user.admin?, params: p).submit
+        r = described_class.new(role: user.role, params: p).submit
         expect(r[:redirect_object]).to be_nil
       end
       it 'returns the form object' do
         new_draw = create(:draw)
         param_hash = { draw_id: new_draw.id.to_s, group_ids: [[group.id]] }
         p = instance_spy('ActionController::Parameters', to_h: param_hash)
-        r = described_class.new(admin: user.admin?, params: p).submit
+        r = described_class.new(role: user.role, params: p).submit
         expect(r[:form_object]).to be_instance_of(described_class)
       end
+    end
+  end
+
+  shared_examples 'clipping restricted' do
+    it 'sets the redirect object to nil' do
+      params = restricted_draw_setup
+      r = described_class.new(role: user.role, params: params).submit
+      expect(r[:redirect_object]).to be_nil
+    end
+
+    it 'returns the form object' do
+      params = restricted_draw_setup
+      r = described_class.new(role: user.role, params: params).submit
+      expect(r[:form_object]).to be_instance_of(described_class)
+    end
+
+    def restricted_draw_setup
+      restricted_draw = create(:draw, status: 'group_formation',
+                                      restrict_clipping_group_size: true)
+      quad = create(:locked_group, :defined_by_draw,
+                    draw: restricted_draw, size: '4')
+      trip = create(:locked_group, :defined_by_draw,
+                    draw: restricted_draw, size: '1')
+      { draw_id: restricted_draw.id.to_s,
+        group_ids: [quad.id.to_s, trip.id.to_s] }
     end
   end
 
@@ -161,11 +180,19 @@ RSpec.describe NewClipForm do
 
     it_behaves_like 'success' do
       let(:size) { 2 }
-      let(:add_self) { nil }
     end
 
     it_behaves_like 'failure' do
       let(:invalid_groups) { [group_ids.first] }
+    end
+
+    it_behaves_like 'clipping restricted'
+
+    it 'creates a fully confirmed clip' do
+      param_hash = { draw_id: draw.id.to_s, group_ids: group_ids }
+      p = instance_spy('ActionController::Parameters', to_h: param_hash)
+      described_class.new(role: user.role, params: p).submit
+      expect(ClipMembership.all.map(&:confirmed)).to match_array([true, true])
     end
   end
 
@@ -177,19 +204,33 @@ RSpec.describe NewClipForm do
 
       it_behaves_like 'success' do
         let(:size) { 3 }
-        let(:add_self) { '1' }
       end
     end
 
     context 'without including themselves in the group' do
       it_behaves_like 'success' do
         let(:size) { 2 }
-        let(:add_self) { '0' }
       end
     end
 
     it_behaves_like 'failure' do
       let(:invalid_groups) { [group_ids.first] }
+    end
+
+    it_behaves_like 'clipping restricted'
+
+    it "doesn't automatically includes the rep's group in the clip" do
+      param_hash = { draw_id: draw.id.to_s, group_ids: group_ids }
+      p = instance_spy('ActionController::Parameters', to_h: param_hash)
+      described_class.new(role: user.role, params: p).submit
+      expect(ClipMembership.all.count).to eq(2)
+    end
+
+    it 'creates a fully unconfirmed clip' do
+      param_hash = { draw_id: draw.id.to_s, group_ids: group_ids }
+      p = instance_spy('ActionController::Parameters', to_h: param_hash)
+      described_class.new(role: user.role, params: p).submit
+      expect(ClipMembership.all.map(&:confirmed)).to match_array([false, false])
     end
   end
 
@@ -199,11 +240,32 @@ RSpec.describe NewClipForm do
 
     it_behaves_like 'success' do
       let(:size) { 3 }
-      let(:add_self) { '1' }
     end
 
     it_behaves_like 'failure' do
       let(:invalid_groups) { [''] }
+    end
+
+    it 'automatically includes the user\'s group in the clip membership' do
+      param_hash = { draw_id: draw.id.to_s, group_ids: group_ids }
+      p = instance_spy('ActionController::Parameters', to_h: param_hash)
+      described_class.new(role: user.role, params: p).submit
+      expect(ClipMembership.all.count).to eq(3)
+    end
+
+    it 'creates a partially confirmed clip' do
+      param_hash = { draw_id: draw.id.to_s, group_ids: group_ids }
+      p = instance_spy('ActionController::Parameters', to_h: param_hash)
+      described_class.new(role: user.role, params: p).submit
+      expect(ClipMembership.all.map(&:confirmed))
+        .to match_array([true, false, false])
+    end
+
+    it 'only confirms their group\'s membership in the clip' do
+      param_hash = { draw_id: draw.id.to_s, group_ids: group_ids }
+      p = instance_spy('ActionController::Parameters', to_h: param_hash)
+      described_class.new(role: user.role, params: p).submit
+      expect(user.group.clip_membership.confirmed).to be(true)
     end
   end
 end