+++ spec/policies/clip_policy_spec.rb
@@ -82,7 +82,7 @@ RSpec.describe ClipPolicy do
       end
     end
 
-    permissions :create_as_rep?, :edit?, :update?, :destroy? do
+    permissions :edit?, :update?, :destroy? do
       it { is_expected.not_to permit(user, clip) }
     end
   end
@@ -92,49 +92,6 @@ RSpec.describe ClipPolicy do
 
     permissions :show?, :create? do
       it { is_expected.to permit(user, clip) }
-    end
-
-    permissions :create_as_rep? do
-      let(:group) { instance_spy('Group', present?: true) }
-
-      context 'not in a group' do
-        before { allow(user).to receive(:group).and_return(nil) }
-        it { is_expected.not_to permit(user, clip) }
-      end
-      context 'in a group' do
-        let(:draw) { instance_spy('Draw') }
-
-        before do
-          allow(user).to receive(:group).and_return(group)
-          allow(clip).to receive(:draw).and_return(draw)
-        end
-
-        context 'but the draw is not in a group-formation stage' do
-          before do
-            allow(draw).to receive(:group_formation?).and_return(false)
-          end
-          it { is_expected.not_to permit(user, clip) }
-        end
-
-        context 'and is in a group-formation draw' do
-          before do
-            allow(draw).to receive(:group_formation?).and_return(true)
-          end
-
-          context 'but the draw does not allow for clipping' do
-            before do
-              allow(draw).to receive(:allow_clipping).and_return(false)
-            end
-            it { is_expected.not_to permit(user, clip) }
-          end
-          context 'and the draw allows for clipping' do
-            before do
-              allow(draw).to receive(:allow_clipping).and_return(true)
-            end
-            it { is_expected.to permit(user, clip) }
-          end
-        end
-      end
     end
 
     permissions :edit?, :update?, :destroy? do
@@ -199,10 +156,6 @@ RSpec.describe ClipPolicy do
         end
         it { is_expected.to permit(user, clip) }
       end
-    end
-
-    permissions :create_as_rep? do
-      it { is_expected.not_to permit(user, clip) }
     end
   end
 end